pyproject.toml
[tool.poetry]
name = "flowers"
version = "0.1.0"
description = ""
authors = ["Casey Chance <caseyjchance@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "*"
pdfquery = "^0.4.3"
camelot-py = {version = "0.9.0", extras = ["base"]}
opencv-python = "^4.7.0.72"
pypdf2 = "2.12.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipdb = "^0.13.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
